pyproject.toml
[project]
name = "hitl"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "aiosqlite>=0.22.1",
    "beautifulsoup4>=4.14.3",
    "ddgs>=9.14.1",
    "duckduckgo-search>=8.1.1",
    "glow>=0.15.10.11",
    "grandalf>=0.8",
    "langchain>=1.2.17",
    "langchain-cerebras>=0.8.2",
    "langchain-community>=0.4.1",
    "langchain-groq>=1.1.2",
    "langchain-openai>=1.2.1",
    "langchain-tavily>=0.2.18",
    "langgraph>=1.1.10",
    "langgraph-checkpoint-sqlite>=3.0.3",
    "langsmith[claude-agent-sdk]>=0.8.0",
    "lxml>=6.1.0",
    "matplotlib>=3.10.9",
    "md2term>=1.0.2",
    "mdrender>=0.1.0",
    "numpy>=2.4.4",
    "pandas>=3.0.2",
    "pydantic>=2.13.3",
    "python-dotenv>=1.2.2",
    "requests>=2.33.1",
    "rich>=15.0.0",
    "scipy>=1.17.1",
    "seaborn>=0.13.2",
    "textual>=8.2.5",
    "yfinance>=1.3.0",
]
